testdrive: reorganize according to Rust conventions
Move all source files into an src directory, and also introduce a format
module for the growing list of formats we support.
