+++ src/testdrive/Cargo.toml
@@ -5,13 +5,6 @@ version = "0.1.0"
 edition = "2018"
 publish = false
 
-[lib]
-path = "lib.rs"
-
-[[bin]]
-name = "testdrive"
-path = "main.rs"
-
 [dependencies]
 atty = "0.2"
 avro = { path = "../avro" }
