+++ src/testdrive/src/action/avro_ocf.rs
@@ -52,7 +52,7 @@ impl Action for WriteAction {
             .map_err(|e| format!("parsing avro schema: {}", e))?;
         let mut writer = Writer::new(&schema, &mut file);
         for record in &self.records {
-            let record = crate::avro::json_to_avro(
+            let record = crate::format::avro::json_to_avro(
                 &serde_json::from_str(record).map_err(|e| format!("parsing avro datum: {}", e))?,
                 &schema,
             )?;
