+++ src/testdrive/src/format/protobuf/simple.proto
